requirements.txt
python-telegram-bot
telebot
motor
pydantic
pytz
pymongo
python-dotenv
APScheduler
matplotlib
pandas
psutil
aiohttp
pyTelegramBotAPI